adding more permissions to lambda
